Rename v2-bug-notice => v2-bug-updates
