+++ src/Router.tsx
@@ -3,7 +3,7 @@ import PrivacyPolicy from 'pages/privacy'
 import Projects from 'pages/projects'
 import Loading from 'components/Loading'
 import V1Dashboard from 'pages/p'
-import V2BugUpdates from 'pages/v2-bug-notice'
+import V2BugUpdates from 'pages/v2-bug-updates'
 
 import { fathom } from 'lib/fathom'
 import { V1CurrencyProvider } from 'providers/v1/V1CurrencyProvider'
